:sparkles: NFT 관련 CRU
NFT CRU 기능 구현 및 테스트 완료 + 추후 제약사항 리팩토링 필요

+++ backend/FANFTASY/src/main/java/com/a306/fanftasy/domain/nft/dto/NFTTradeDTO.java
@@ -19,8 +19,8 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 @Builder
 public class NFTTradeDTO {
-    int nftId;
-    int buyerId;
+    long nftId;
+    long buyerId;
     @DateTimeFormat(pattern = "yyyy-MM-dd'T'HH:mm:ss")
     LocalDateTime transactionTime;
